Move guarded containers into their own module
